pyproject.toml
[project]
name = ""
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit",
    "livekit-api",
    "python-dotenv"
]